+++ src/index.js
@@ -1,3 +1,4 @@
 export * from "./WebCrypt.js";
 export * from "./WebCryptAsym.js";
 export * from "./WebCryptPQC.js";
+export { default as TimingSafeHelper } from "./TimingSafeHelper.js";